requirements.txt
javabridge==1.0.18
python-weka-wrapper3==0.1.12
numpy==1.18.3
pandas==1.0.3
pandas-datareader==0.8.1
pandas-gbq==0.11.0
pandas-profiling==1.4.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
sklearn==0.0
sklearn-pandas==1.8.0
